+++ CHANGELOG.md
@@ -6,6 +6,10 @@ in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [unreleased]
+### Added
+- Adding support for shuffling a list of strings.
+
 ## [0.8.1] - 2019-01-01
 ### Changed
 - Update help on the PercentLikely command